refactor: reorganize and update imports for `LogoutUserService` and `LoginUserService`
- Moved `DTO`, `Command`, and `Validator` files to appropriate `Application` and `Domain` subdirectories.
- Updated import paths across affected services and handlers.
