+++ user-service/src/Application/Services/Concrete/LoginUserService.ts
@@ -1,10 +1,10 @@
 import {BaseService} from "../Interfaces/BaseService.js";
-import {UserSessionDTO} from "../../../Domain/DTO/Command/UserSessionDTO.js";
-import {LoginSessionCommandHandler} from "../../Command/Handlers/LoginSessionCommandHandler.js";
-import {LoginSessionCommandValidator} from "../../Command/Validators/LoginSessionCommandValidator.js";
+import {UserSessionDTO} from "../../DTO/ToCommand/UserSessionDTO.js";
+import {LoginSessionCommandHandler} from "../../../Domain/Command/Handlers/LoginSessionCommandHandler.js";
+import {LoginSessionCommandValidator} from "../../../Domain/Command/Validators/LoginSessionCommandValidator.js";
 import {UserRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/UserRepository.js";
 import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
-import {UserSessionCommand} from "../../Command/CommandObject/UserSessionCommand.js";
+import {UserSessionCommand} from "../../../Domain/Command/CommandObject/UserSessionCommand.js";
 import {ValidationException} from "../../../Shared/Errors/ValidationException.js";
 import {PrismaClientKnownRequestError} from "@prisma/client/runtime/library";
 import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog.js";
