+++ user-service/src/Application/Services/Concrete/LogoutUserService.ts
@@ -1,10 +1,10 @@
-import {UserSessionDTO} from "../../../Domain/DTO/Command/UserSessionDTO.js";
+import {UserSessionDTO} from "../../DTO/ToCommand/UserSessionDTO.js";
 import {UserRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/UserRepository.js";
 import {NotificationError} from "../../../Shared/Errors/NotificationError.js";
 import {BaseService} from "../Interfaces/BaseService.js";
-import {UserSessionCommand} from "../../Command/CommandObject/UserSessionCommand.js";
-import {LogoutSessionCommandHandler} from "../../Command/Handlers/LogoutSessionCommandHandler.js";
-import {LogoutSessionCommandValidator} from "../../Command/Validators/LogoutSessionCommandValidator.js";
+import {UserSessionCommand} from "../../../Domain/Command/CommandObject/UserSessionCommand.js";
+import {LogoutSessionCommandHandler} from "../../../Domain/Command/Handlers/LogoutSessionCommandHandler.js";
+import {LogoutSessionCommandValidator} from "../../../Domain/Command/Validators/LogoutSessionCommandValidator.js";
 import {ValidationException} from "../../../Shared/Errors/ValidationException.js";
 import {PrismaClientKnownRequestError} from "@prisma/client/runtime/library";
 import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog.js";
