+++ user-service/src/Domain/Command/Handlers/LogoutSessionCommandHandler.ts
